Criação de validações, filter para validar os dtos

+++ Backend/SalesWeb/SalesWeb.Domain/Entities/Seller.cs
@@ -9,7 +9,7 @@ namespace SalesWeb.Domain.Entities
         public string Name { get; set; }
         public string Email { get; set; }
         public DateTime BirthDate { get; set; }
-        public double BaseSalary { get; set; }
+        public decimal BaseSalary { get; set; }
         public Department Department { get; set; }
         public int DepartmentId { get; set; }
         //public ICollection<SalesRecord> Sales { get; set; } = new List<SalesRecord>();
@@ -17,7 +17,7 @@ namespace SalesWeb.Domain.Entities
 
         public Seller() { }
 
-        public Seller(string name, string email, DateTime birthDate, double baseSalary, Department department)
+        public Seller(string name, string email, DateTime birthDate, decimal baseSalary, Department department)
         {
             Name = name;
             Email = email;
