+++ Backend/SalesWeb/SalesWeb/DTO/SellerDTO.cs
@@ -7,7 +7,7 @@ namespace SalesWeb.DTO
         public string Name { get; set; }
         public string Email { get; set; }
         public DateTime BirthDate { get; set; }
-        public double BaseSalary { get; set; }
+        public decimal BaseSalary { get; set; }
         public int DepartmentId { get; set; }
     }
 }
